requirements.txt
streamlit
yfinance
plotly
pandas_ta
pandas
newsapi-python
textblob
deep_translator